Factoring out ret2spec_common.cc
